+++ frontend/pages/index.tsx
@@ -4,20 +4,17 @@ import { GetStaticProps, NextPage } from "next";
 import { ExchangeRate, getRates } from "../lib/exchangeRate";
 import { Status, getLatestStatus } from "../lib/status";
 import { amountFormatter } from "../lib/formatter";
+import Header from "../components/Header";
 import Input from "../components/Input";
 import Footer from "../components/Footer";
-import RateRow from "../components/RateRow";
-import ThemeSwitch from "../components/ThemeSwitch";
-import dayjs from "dayjs";
-import relativeTime from "dayjs/plugin/utc";
-import utc from "dayjs/plugin/relativeTime";
+import MidRate from "../components/MidRate";
+import RatesTable from "../components/RatesTable";
+import LastUpdated from "../components/LastUpdated";
 
 type HomePageProps = {
   rates: ExchangeRate[];
   status: Status;
 };
-
-const MID_RATE_TAG = "MidRate";
 
 export const getStaticProps: GetStaticProps<HomePageProps> = async () => {
   const rates = await getRates();
@@ -29,7 +26,7 @@ export const getStaticProps: GetStaticProps<HomePageProps> = async () => {
 };
 
 const Home: NextPage<HomePageProps> = ({ rates, status }) => {
-  const midRateData = rates.find((r) => r.source === MID_RATE_TAG);
+  const midRateData = rates.find((r) => r.source === "MidRate");
   if (midRateData == null) {
     throw new Error("No midrate found");
   }
@@ -71,30 +68,6 @@ const Home: NextPage<HomePageProps> = ({ rates, status }) => {
     }
   };
 
-  // Sort by best rate
-  const sortedrates = [...rates].sort((a, b) => {
-    if (a.rate > b.rate) {
-      return -1;
-    }
-    if (a.rate < b.rate) {
-      return 1;
-    }
-    return 0;
-  });
-
-  dayjs.extend(relativeTime);
-  dayjs.extend(utc);
-
-  const updatedOnUtc = dayjs.utc(status.updatedOn);
-  const utcNow = new Date().toUTCString();
-  const lastUpdated = updatedOnUtc.from(utcNow);
-
-  const ratesRows = sortedrates
-    .filter((rate) => rate.source !== MID_RATE_TAG)
-    .map((rate) => {
-      return <RateRow key={rate._id} rateData={rate} amount={sgdAmount} />;
-    });
-
   return (
     <>
       <Head>
@@ -107,25 +80,8 @@ const Home: NextPage<HomePageProps> = ({ rates, status }) => {
         <link rel="icon" href="/peso.png" />
       </Head>
       <div className="flex flex-col items-center min-w-full min-h-screen mx-auto pt-4 px-4 dark:bg-gray-900 dark:text-gray-100">
-        <header>
-          <div className="flex justify-between w-96 px-2">
-            <span className="text-sm hover:underline decoration-yellow-400 decoration-2">
-              <a href="mailto:hey@sgdtopeso.com">Say Hi!</a>
-            </span>
-            <span>
-              <ThemeSwitch />
-            </span>
-          </div>
-          <div className="text-center text-3xl mt-8">
-            <h1 className="font-bold underline">SGD to PHP Rates</h1>
-          </div>
-        </header>
-        <div className="text-center font-semibold mt-8 mb-4 h-14">
-          <p className="text-slate-400">MARKET RATE: 1 SGD =</p>
-          <p className="text-xl hover:text-2xl font-bold duration-200">
-            {midRate} Philippine Pesos
-          </p>
-        </div>
+        <Header />
+        <MidRate midRate={midRate} />
         <div className="flex justify-center">
           <div className="flex flex-col items-center md:flex-row space-y-2 md:space-y-0">
             <div className="px-4 md:my-0">
@@ -162,25 +118,9 @@ const Home: NextPage<HomePageProps> = ({ rates, status }) => {
             </div>
           </div>
         </div>
-        <div className="mt-10">
-          <p className="text-sm">Last updated {lastUpdated}</p>
-        </div>
-        <div className="flex justify-center mt-4 mb-8">
-          <table className="table-fixed text-left shadow-sm">
-            <thead>
-              <tr className="text-xs uppercase text-gray-500 bg-gray-200 dark:bg-gray-700 dark:text-gray-100">
-                <th className="pl-4 py-2">Send with</th>
-                <th className="px-4">Rate</th>
-                <th className="px-4">Fees</th>
-                <th className="px-4 w-40">You get</th>
-              </tr>
-            </thead>
-            <tbody>{ratesRows}</tbody>
-          </table>
-        </div>
-        <div>
-          <Footer />
-        </div>
+        <LastUpdated status={status} />
+        <RatesTable rates={rates} sgdAmount={sgdAmount} />
+        <Footer />
       </div>
     </>
   );
